src/use-cases/fetch-user-check-ins-history.ts
import type { CheckIn } from '@prisma/client'

import { CheckInsRepository } from '@/repositotories/check-ins-repository'
import { UsersRepository } from '@/repositotories/users-repository'
import { ResourceNotFoundError } from './errors/resource-not-found-error'

interface FetchUserCheckInsHistoryUseCaseRequest {
  userId: string
  page: number
}

interface FetchUserCheckInsHistoryUseCaseResponse {
  checkIns: CheckIn[]
}

export class FetchUserCheckInsHistoryUseCase {
  constructor(
    private checkInsRepository: CheckInsRepository,
    private userRepository: UsersRepository,
  ) {}

  async execute({
    userId,
    page,
  }: FetchUserCheckInsHistoryUseCaseRequest): Promise<FetchUserCheckInsHistoryUseCaseResponse> {
    const user = await this.userRepository.findById(userId)

    if (!user) throw new ResourceNotFoundError()

    const checkIns = await this.checkInsRepository.findManyByUserId(
      userId,
      page,
    )

    return { checkIns }
  }
}
